Rename focus state to focusPending
This commit makes the focus part of the state more precise. Namely, `focus` implies that we always _track_ focus, while `focusPending` hints that we merely requesed a refocusing. Tracking focus of the input element is actually a much harder task and not necessary for the usability (*i.e.*, we also want to allow the user to _blur_ on purpose).