Fix dependency initialization as promise (#41)
It takes a while for the browser to initialize speech synthesis. Therefore, `speechSynthesis.getVoices()` returned unexpectedly sometimes an empty array.

This commit fixes the problem by refactoring the initialization of the dependencies into a promise and showing the progress bar during the initialization.

Additionally, the commit introduces an error boundary to display any errors which might occur.