Fix voices detection on Android (#48)
This commit makes a loop that checks repeatedly for `speechSynthesis.getVoices()` and rejects after certain number of retries. This seems to be the best approach for the browsers running on Android.

BCP 47 language tags use underscore (`_`) instead of dash (`-`) on Android, so this had to be fixed. The `lang` property also had to be set on `SpeechSynthesisUtterance` in order to effectuate the language change.